# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.AuthMethod:
    enum:
    - sms
    - qr
    type: string
    x-enum-varnames:
    - AuthMethodSMS
    - AuthMethodQR
...
  domain.CreateSessionRequest:
    properties:
      api_hash:
        type: string
      api_id:
        type: integer
      auth_method:
        $ref: '#/definitions/domain.AuthMethod'
      phone:
        type: string
      session_name:
        type: string
    required:
    - api_hash
    - api_id
    type: object
...
  domain.SessionStatus:
    enum:
    - pending
    - code_sent
    - password_required
    - authenticated
    - failed
    type: string
    x-enum-varnames:
    - SessionPending
    - SessionCodeSent
    - SessionPasswordRequired
    - SessionAuthenticated
    - SessionFailed
  domain.TelegramSession:
    properties:
      api_id:
        type: integer
      auth_state:
        $ref: '#/definitions/domain.SessionStatus'
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      phone_number:
        type: string
      session_name:
        type: string
      telegram_user_id:
        type: integer
      telegram_username:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  domain.VerifyCodeRequest:
    properties:
      code:
        maxLength: 6
        minLength: 5
        type: string
    required:
    - code
    type: object
...
  domain.WebhookConfig:
    properties:
      created_at:
        type: string
      events:
        description: Tipos de eventos a enviar
        items:
          type: string
        type: array
      id:
        type: string
      is_active:
        type: boolean
      last_error:
        type: string
      last_error_at:
        type: string
      max_retries:
        type: integer
      secret:
        description: Para firmar requests
        type: string
      session_id:
        type: string
      timeout_ms:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
  domain.WebhookCreateRequest:
    properties:
      events:
        example:
        - message.new
        - message.edit
        items:
          type: string
        type: array
      max_retries:
        example: 3
        type: integer
      secret:
        example: mi_secret_123
        type: string
      timeout_ms:
        example: 5000
        type: integer
      url:
        example: https://mi-servidor.com/webhook
        type: string
    required:
    - url
    type: object
  domain.WebhookResponse:
    properties:
      events:
        items:
          type: string
        type: array
      id:
        type: string
      is_active:
        type: boolean
      session_id:
        type: string
      url:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  handler.MetaResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorResponse'
      meta:
        $ref: '#/definitions/handler.MetaResponse'
      success:
        type: boolean
    type: object
host: localhost:7789
info:
  contact: {}
  description: API REST para gestionar múltiples sesiones de Telegram via MTProto
  title: Telegram API
  version: 0.1.0
paths:
...
  /pool/status:
    get:
      description: Retorna información de sesiones activas escuchando
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Estado del pool
      tags:
      - Webhooks
  /sessions:
    get:
      description: Retorna todas las sesiones Telegram del usuario
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.TelegramSession'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Listar sesiones
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Inicia autenticación con Telegram (SMS o QR). Para QR, el sistema
        escucha automáticamente en background.
      parameters:
      - description: Credenciales Telegram
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Crear sesión Telegram
      tags:
      - Sessions
  /sessions/{id}:
    delete:
      description: Elimina una sesión Telegram
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Eliminar sesión
      tags:
      - Sessions
    get:
      description: Retorna detalle de una sesión. Use para verificar si QR fue escaneado
        (is_active=true).
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.TelegramSession'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Obtener sesión
      tags:
      - Sessions
...
  /sessions/{id}/verify:
    post:
      consumes:
      - application/json
      description: Completa autenticación con el código recibido por SMS
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Código SMS
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.VerifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.TelegramSession'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Verificar código SMS
      tags:
      - Sessions
  /sessions/{id}/webhook:
    delete:
      description: Elimina configuración de webhook
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Eliminar webhook
      tags:
      - Webhooks
    get:
      description: Retorna configuración actual del webhook
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.WebhookConfig'
              type: object
      security:
      - BearerAuth: []
      summary: Obtener webhook
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Configura URL de webhook para recibir eventos de la sesión
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Configuración
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.WebhookCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.WebhookResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Configurar webhook
      tags:
      - Webhooks
  /sessions/{id}/webhook/start:
    post:
      description: Inicia la escucha de eventos de Telegram para esta sesión
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Iniciar escucha
      tags:
      - Webhooks
  /sessions/{id}/webhook/stop:
    post:
      description: Detiene la escucha de eventos para esta sesión
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Detener escucha
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"